requirements.txt
Flask==2.1.0
openai==1.3.0
python-dotenv==0.19.2
nltk==3.6.5
gunicorn==20.1.0
